+++ src/server/webhook_test.go
@@ -149,6 +149,7 @@ func TestServeHTTP(t *testing.T) {
 			NewEnvVarMutator(clusterName),
 			NewSidecarMutator(clusterName, makeConfigMapRetriever("default", configName, map[string]string{"config.yaml": integrationConfig})),
 		},
+		IgnoreNamespaces: []string{metav1.NamespaceSystem, metav1.NamespacePublic},
 	}
 
 	server := httptest.NewServer(whsvr)
@@ -176,6 +177,103 @@ func TestServeHTTP(t *testing.T) {
 		})
 	}
 
+}
+
+func TestServeHTTPIgnoreNamespaces(t *testing.T) {
+	expectedEnvVarsPatchForValidBody := loadTestData(t, "expectedEnvVarsAdmissionReviewPatch.json")
+	configName := "my-config"
+
+	patchTypeForValidBody := v1beta1.PatchTypeJSONPatch
+	cases := []struct {
+		name                      string
+		requestBody               []byte
+		contentType               string
+		expectedStatusCode        int
+		expectedBodyWhenHTTPError string
+		expectedAdmissionReview   v1beta1.AdmissionReview
+	}{
+		{
+			name:               "mutation applied - namespace kube-system",
+			requestBody:        makeTestData(t, "kube-system", nil),
+			contentType:        "application/json",
+			expectedStatusCode: http.StatusOK,
+			expectedAdmissionReview: v1beta1.AdmissionReview{
+				Response: &v1beta1.AdmissionResponse{
+					UID:       types.UID(1),
+					Allowed:   true,
+					Result:    nil,
+					Patch:     expectedEnvVarsPatchForValidBody,
+					PatchType: &patchTypeForValidBody,
+				},
+			},
+		},
+		{
+			name:               "mutation applied - namespace kube-public",
+			requestBody:        makeTestData(t, "kube-public", nil),
+			contentType:        "application/json",
+			expectedStatusCode: http.StatusOK,
+			expectedAdmissionReview: v1beta1.AdmissionReview{
+				Response: &v1beta1.AdmissionResponse{
+					UID:       types.UID(1),
+					Allowed:   true,
+					Result:    nil,
+					Patch:     expectedEnvVarsPatchForValidBody,
+					PatchType: &patchTypeForValidBody,
+				},
+			},
+		},
+		{
+			name:               "mutation not applied - namespace testing ignored",
+			requestBody:        makeTestData(t, "testing", nil),
+			contentType:        "application/json",
+			expectedStatusCode: http.StatusOK,
+			expectedAdmissionReview: v1beta1.AdmissionReview{
+				Response: &v1beta1.AdmissionResponse{
+					UID:       types.UID(1),
+					Allowed:   true,
+					Result:    nil,
+					Patch:     nil,
+					PatchType: nil,
+				},
+			},
+		},
+	}
+
+	clusterName := "foobar"
+	whsvr := &Webhook{
+		ClusterName: clusterName,
+		Server:      &http.Server{},
+		Mutators: []podMutator{
+			NewEnvVarMutator(clusterName),
+			NewSidecarMutator(clusterName, makeConfigMapRetriever("default", configName, map[string]string{"config.yaml": integrationConfig})),
+		},
+		IgnoreNamespaces: []string{"testing"},
+	}
+
+	server := httptest.NewServer(whsvr)
+	defer server.Close()
+
+	for i, c := range cases {
+		t.Run(fmt.Sprintf("[%d] %s", i, c.name), func(t *testing.T) {
+
+			fmt.Println(c.name)
+			resp, err := http.Post(server.URL, c.contentType, bytes.NewReader(c.requestBody))
+			assert.NoError(t, err)
+			assert.Equal(t, c.expectedStatusCode, resp.StatusCode)
+
+			gotBody, err := ioutil.ReadAll(resp.Body)
+			if err != nil {
+				t.Fatalf("could not read body: %v", err)
+			}
+			var gotReview v1beta1.AdmissionReview
+			if err := json.Unmarshal(gotBody, &gotReview); err != nil {
+				assert.Equal(t, c.expectedBodyWhenHTTPError, string(gotBody))
+				return
+			}
+
+			assert.Equal(t, c.expectedAdmissionReview, gotReview)
+		})
+	}
 }
 
 func Benchmark_EnvVarWebhookPerformance(b *testing.B) {